requirements.txt
dj-static==0.0.6
Django==1.10.5
django-suit==0.2.25
djangorestframework==3.5.4
gunicorn==19.6.0
MySQL-python==1.2.5
wheel==0.24.0
whitenoise==3.3.0
